+++ Classes/Service/CountryService.php
@@ -13,11 +13,18 @@ namespace Brotkrueml\FormCountrySelect\Service;
 
 use Brotkrueml\FormCountrySelect\Domain\Model\FormElements\CountrySelect;
 use Brotkrueml\FormCountrySelect\Event\CountriesModificationEvent;
+use Brotkrueml\FormCountrySelect\Extension;
 use Symfony\Component\Intl\Countries;
+use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\EventDispatcher\EventDispatcher;
 use TYPO3\CMS\Core\Information\Typo3Version;
+use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\SignalSlot\Dispatcher;
+
+if (!Environment::isComposerMode() && !\class_exists(Countries::class)) {
+    @include 'phar://' . ExtensionManagementUtility::extPath(Extension::KEY) . 'Resources/Private/PHP/symfony-intl.phar/vendor/autoload.php';
+}
 
 /**
  * @api