+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Include package symfony/intl for non-composer installations when missing
+
 ## [1.3.0] - 2020-12-25
 
 ### Added